+++ src/tools/rust-analyzer/lib/smol_str/CHANGELOG.md
@@ -5,6 +5,8 @@
 - Optimise `StrExt::to_ascii_lowercase_smolstr`, `StrExt::to_ascii_uppercase_smolstr` 
   ~2x speedup inline, ~4-22x for heap.
 - Optimise `StrExt::to_lowercase_smolstr`, `StrExt::to_uppercase_smolstr` ~2x speedup inline, ~5-50x for heap.
+- Optimise `StrExt::replace_smolstr`, `StrExt::replacen_smolstr` for single ascii replace.
+  ~3x speedup inline, ~1.8x for heap (len=50).
 
 ## 0.3.2 - 2024-10-23